Flyttet funksjonalitet fra pdfSøknadController til søknadKvitteringController
